added tournament delete confirmation

+++ app/assets/javascripts/controllers/tournament_controller.js
@@ -1,10 +1,12 @@
 Gameway.TournamentController = Ember.ObjectController.extend({
   actions: {
     destroy: function(model) {
-      var thisController = this;
-      model.destroyRecord().then(function() {
-        thisController.transitionToRoute('tournaments.index')
-      })
+      if (confirm("Are you sure you want to delete %@?".fmt(model.get('title')))) {
+        var thisController = this;
+        model.destroyRecord().then(function() {
+          thisController.transitionToRoute('tournaments.index')
+        })
+      }
     }
   }
 })
